backend/requirements.txt
boto3==1.34.51
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
reportlab==4.0.9